move post tests into separate file
